+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 7.0.0
+* Porting to flutter_map v 7.0
+* Updated dependencies to the latest version
+* Added offline cache
+
 ## 6.0.1
 * Porting also dragmarker and update the editing part
 * Add macos support for editing maps